Functionality for toggle box extracted to ToggleBox component
