--- client/src/components/shared/Info/Finances.scss
@@ -1,3 +1,0 @@
-.older-finances {
-  margin-top: .5em;
-}
